Implement Viewport methods in Editor
